[ELIZAAI-16](feat): clean up token transfer
