+++ packages/plugin-cosmos/src/actions/transfer/schema.ts
@@ -1,0 +1,8 @@
+import { z } from "zod";
+
+export const cosmosTransferParamsSchema = z.object({
+    chainName: z.string(),
+    symbol: z.string(),
+    amount: z.string(),
+    toAddress: z.string(),
+});
