+++ packages/plugin-cosmos/src/actions/transfer/types.ts
@@ -1,0 +1,4 @@
+import { z } from "zod";
+import { cosmosTransferParamsSchema } from "./schema";
+
+export type CosmosTransferParams = z.infer<typeof cosmosTransferParamsSchema>;
