+++ packages/plugin-cosmos/src/shared/interfaces.ts
@@ -1,0 +1,10 @@
+export interface ICosmosActionService {
+    execute: ((...params: unknown[]) => void) | (() => void);
+}
+
+export interface ICosmosTransaction {
+    from: string;
+    to: string;
+    txHash: string;
+    gasPaid: number;
+}
